rpi3-waveshare-book-clock/requirements.txt
Pillow>=10.0
waveshare-epaper>=1.3.0
